chore: move remaining icons to parents folder
