+++ src/components/career-form/career-file-upload-type.tsx
@@ -11,7 +11,7 @@ import {
 } from 'react-hook-form';
 import { useTranslation } from 'gatsby-plugin-react-i18next';
 import { up } from '../style-utils/breakpoint';
-import { SelectArrow } from '../icons/select-arrow';
+import { SelectArrow } from './icons/select-arrow';
 
 interface CareerFileUploadTypeProps {
   setValue: UseFormSetValue<FieldValues>;
