+++ src/components/career-form/career-form.tsx
@@ -15,9 +15,9 @@ import { CaptionText, TextLink, TextTitle } from '../typography/typography';
 import { Checkbox, Sup } from '../form/controls';
 import { FileUpload } from './career-file-upload';
 import { CareerTextFields } from './career-textfields';
-import { Upload } from '../icons/upload';
+import { Upload } from './icons/upload';
 import styled from 'styled-components';
-import { CheckboxMark } from '../icons/checkbox';
+import { CheckboxMark } from './icons/checkbox';
 import { up } from '../style-utils/breakpoint';
 import { rgba } from 'polished';
 import { Trans, useTranslation } from 'gatsby-plugin-react-i18next';
