+++ src/components/career-form/icons/clear.tsx
