+++ src/components/header/header.tsx
@@ -2,8 +2,8 @@ import React, { useEffect, useState } from 'react';
 import styled from 'styled-components';
 import { up } from '../style-utils/breakpoint';
 import { GRID_GAP_DESKTOP, GRID_GAP_MOBILE } from '../grid/grid';
-import { BurgerMenu } from '../icons/burger-menu';
-import { CloseBurgerMenuIcon } from '../icons/buttons-icons/close-burger-menu';
+import { BurgerMenu } from './icons/burger-menu';
+import { CloseBurgerMenuIcon } from './icons/close-burger-menu';
 import { NavigationFlyout } from './menu-flyout';
 import { Link } from '../links/links';
 import { Swoosh } from '../icons/swoosh';
