--- src/components/icons/byline-arrow.tsx
@@ -1,19 +1,0 @@
-import React from 'react';
-
-export const BylineArrow: React.FC = () => (
-  <svg
-    width="11"
-    height="11"
-    viewBox="0 0 11 11"
-    fill="none"
-    xmlns="http://www.w3.org/2000/svg"
-  >
-    <path
-      d="M5 1H10V6"
-      stroke="#668CFF"
-      strokeWidth="1.5"
-      strokeLinejoin="round"
-    />
-    <path d="M10 1L1 10" stroke="#668CFF" strokeWidth="1.5" />
-  </svg>
-);
