+++ src/pages/blog.tsx
@@ -12,7 +12,7 @@ import { formatDate } from '../components/util/format-date';
 import { getImage, IGatsbyImageData } from 'gatsby-plugin-image';
 import { useI18next, useTranslation } from 'gatsby-plugin-react-i18next';
 import { ButtonText, SendButton } from '../components/form/controls';
-import { RightArrowIcon } from '../components/icons/buttons-icons/right-arrow';
+import { RightArrowIcon } from '../components/career-form/icons/right-arrow';
 import { LocalesQuery } from '../types';
 
 const TOP_POST_COUNT = 2;
